Features/batchify input

+++ segmentation.py
@@ -10,62 +10,43 @@ from sam_functions import get_sam
 import matplotlib.pyplot as plt
 from torchmetrics.classification import BinaryJaccardIndex
 import networkx as nx
+import torch.nn.functional as F
 
 
-def segment_LF(mask_generator, LF):
-    max_segment_num = -1
-    s, t, u, v, c = LF.shape
-    segments = torch.zeros(
-        (
-            s,
-            t,
-            u,
-            v,
-        )
-    )
-    min_mask_area = int(CONFIG["min-mask-area"] * u * v)
-    mask_generator.min_mask_region_area = min_mask_area
-    for i in tqdm(range(LF.shape[0])):
-        for j in range(LF.shape[1]):
-            seg_mask = mask_generator.generate(LF[i][j])
-            seg_mask = sorted(seg_mask, key=(lambda x: x["area"]), reverse=True)
-            seg_mask = filter(lambda mask: mask["area"] >= min_mask_area, seg_mask)
-            segments_ij = torch.stack(
-                [
-                    torch.tensor(mask["segmentation"]).cuda() * (mask_i + 1)
-                    for mask_i, mask in enumerate(seg_mask)
-                ]
-            )
-            segments_ij = (
-                segments_ij.permute(1, 2, 0).argmax(axis=-1) + max_segment_num + 1
-            )
-            max_segment_num = segments_ij.max()
-            segments[i][j] = segments_ij
-    segments = segments.to(torch.int32)
-    return segments.cuda()
-
-
-def get_segments_metric(segments, i, j, metric=BinaryJaccardIndex().cuda()):
+def calculate_segments_metric(segments, embeddings, i, j):
     mask_i = (segments == i).to(torch.int32)
     mask_j = (segments == j).to(torch.int32)
-    segments_i = mask_i.sum(axis=(0, 1))
-    segments_j = mask_j.sum(axis=(0, 1))
-    metric_val = metric(segments_i, segments_j)
-    return metric_val.item()
+    if CONFIG["edge-metric"] in ["cosine", "combination"]:
+        embs_i = (embeddings * (segments == i)[:, :, :, :, None].to(torch.int32)).mean(
+            axis=(0, 1, 2, 3)
+        )
+        embs_j = (embeddings * (segments == j)[:, :, :, :, None].to(torch.int32)).mean(
+            axis=(0, 1, 2, 3)
+        )
+        cosine_sim = F.cosine_similarity(embs_i[None], embs_j[None])[0].item()
+        if CONFIG["edge-metric"] == "cosine":
+            return cosine_sim
+    if CONFIG["edge-metric"] in ["iou", "combination"]:
+        iou_metric = BinaryJaccardIndex().cuda()
+        segments_i = mask_i.sum(axis=(0, 1))
+        segments_j = mask_j.sum(axis=(0, 1))
+        iou = iou_metric(segments_i, segments_j).item()
+        if CONFIG["edge-metric"] == "iou":
+            return iou
+    return CONFIG["cosine-weight"] * cosine_sim + (1 - CONFIG["cosine-weight"]) * iou
 
 
-def get_merge_segments_mapping(segments):
-    metric = BinaryJaccardIndex().cuda()
+def get_merge_segments_mapping(segments, embeddings):
     s, t = segments.shape[0], segments.shape[1]
-    central_inds = torch.unique(segments[s // 2][t // 2].reshape(-1))
+    central_inds = torch.unique(segments[s // 2][t // 2].reshape(-1))[1:]
     graphs = []
     mapping = {}
-    for s_ in range(s):
+    for s_ in tqdm(range(s)):
         for t_ in range(t):
             if s_ == s // 2 and t_ == t // 2:
                 continue
             graph = nx.Graph()
-            segments_st = torch.unique(segments[s_][t_].reshape(-1))
+            segments_st = torch.unique(segments[s_][t_].reshape(-1))[1:]
             top_nodes = segments_st.cpu().detach().numpy()
             bottom_nodes = central_inds.cpu().detach().numpy()
             graph.add_nodes_from(
@@ -78,7 +59,9 @@ def get_merge_segments_mapping(segments):
             )
             for i in central_inds:
                 for j in segments_st:
-                    metric_val = get_segments_metric(segments, i, j, metric)
+                    metric_val = calculate_segments_metric(segments, embeddings, i, j)
+                    if metric_val < CONFIG["metric-threshold"]:
+                        metric_val = 0.0
                     graph.add_edges_from(
                         [(j.item(), i.item())], weight=(1 - metric_val)
                     )
@@ -93,47 +76,57 @@ def get_merge_segments_mapping(segments):
     return mapping
 
 
-def visualize_segments(segments, LF, st_border=None, filename=None):
-    if st_border:
-        LF = LF[st_border:-st_border, st_border:-st_border]
-        segments = segments[st_border:-st_border, st_border:-st_border]
-    s, t, u, v, c = LF.shape
-    segments = np.transpose(segments, (0, 2, 1, 3)).reshape(s * u, t * v)
-    LF = np.transpose(LF, (0, 2, 1, 3, 4)).reshape(s * u, t * v, c)
-    vis = imgviz.label2rgb(
-        label=segments,
-        image=LF,
-        colormap=imgviz.label_colormap(segments.max() + 1),
-    )
-    plt.imshow(vis)
-    if filename:
-        plt.savefig(filename)
-    plt.close()
+def post_process_segments(segments):
+    u, v = segments.shape[-2:]
+    result_segments = []
+    min_mask_area = int(CONFIG["min-mask-area-final"] * u * v)
+    for i in np.unique(segments)[1:]:
+        seg_i = segments == i
+        if seg_i.sum(axis=(2, 3)).mean() >= min_mask_area:
+            result_segments.append(seg_i)
+    return result_segments
 
 
 def main(
     LF_dir,
     segments_filename=CONFIG["segments-filename"],
+    embeddings_filename=CONFIG["embeddings-filename"],
     merged_filename=CONFIG["merged-filename"],
     segments_checkpoint=CONFIG["sam-segments-checkpoint"],
+    merged_checkpoint=CONFIG["merged-checkpoint"],
     vis_filename=CONFIG["vis-filename"],
 ):
-    LF = get_LF(LF_dir)
-    mask_generator = get_sam()
+    from scipy.io import loadmat
+
+    # LF = get_LF(LF_dir)
+    LF = loadmat("lego_128.mat")["LF"].astype(np.int32)[1:-1, 1:-1]
+    simple_sam = get_sam()
     if segments_checkpoint and os.path.exists(segments_filename):
         segments = torch.load(segments_filename).cuda()
+        embeddings = torch.load(embeddings_filename).cuda()
     else:
-        segments = segment_LF(mask_generator, LF)
+        segments, embeddings = simple_sam.segment_LF(LF)
         torch.save(segments, segments_filename)
-    mapping = get_merge_segments_mapping(segments)
-    segments = segments.cpu().numpy()
-    segments = np.vectorize(lambda x: mapping.get(x, x))(segments)
-    torch.save(segments, merged_filename)
+        torch.save(embeddings, embeddings_filename)
+    if merged_checkpoint and os.path.exists(merged_filename):
+        segments = torch.load(merged_filename)
+    else:
+        mapping = get_merge_segments_mapping(segments, embeddings)
+        segments = segments.cpu().numpy()
+        segments = np.vectorize(lambda x: mapping.get(x, x))(segments)
+        torch.save(segments, merged_filename)
     visualize_segments(
         segments,
         LF,
         filename=vis_filename,
     )
+    segments = post_process_segments(segments)
+    for i, segment in enumerate(segments):
+        visualize_segments(
+            segment.astype(np.uint32),
+            LF * (segment).astype(np.int32)[:, :, :, :, None],
+            filename=f"imgs/{str(i).zfill(3)}.png",
+        )
     return segments
